Moved url-state to lib/url-state and created new file for initApp and initRoute
