+++ src/index.js
@@ -9,6 +9,9 @@ import ReactDOM from 'react-dom'
 import './index.css'
 import App from './components/App'
 import registerServiceWorker from './registerServiceWorker'
+import { initApp } from './init'
+
+initApp()
 
 ReactDOM.render(<App />, document.getElementById('root'))
 registerServiceWorker()
